pss-drift-detection/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pss-drift-detection"
version = "0.1.0"
description = "PSS Drift Detection Integration for Neo4j Agent Ecosystem"
requires-python = ">=3.10"
dependencies = [
    "neo4j>=5.11",
    "numpy>=1.24",
    "sentence-transformers>=2.2",
    "openai>=1.0",
    "python-dotenv>=1.0",
    "requests>=2.28",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.1",
    "testcontainers[neo4j]>=3.7",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
markers = [
    "integration: marks tests requiring a running Neo4j instance",
    "e2e: marks end-to-end tests",
]

[tool.coverage.run]
source = ["src"]
omit = ["tests/*", "benchmarks/*"]

[tool.coverage.report]
fail_under = 80
show_missing = true
